feat: realise book ui
