+++ src/app/app.component.html
@@ -1,12 +1,10 @@
 <h1 #appTitle>{{title}}</h1>
 <p>{{testTitle}}</p>
-<app-book
-  [bookShop]="_bookShop"
-  (buy)="onBuy($event)"
-></app-book>
 
-<app-cart
+<app-book-list
   class="flex"
-  [cart]="cart"
-  (removeFromCartAction)="setBackIntoShop($event)"
-></app-cart>
+></app-book-list>
+
+<app-cart-list>
+  <app-cart-item></app-cart-item>
+</app-cart-list>
