+++ src/app/components/book/components/book-item/book-item.component.ts
@@ -1,4 +1,5 @@
-import { Component, OnInit } from '@angular/core';
+import {Component, Input, OnInit, EventEmitter, Output} from '@angular/core';
+import {BookModel} from '../../models/book-model';
 
 @Component({
   selector: 'app-book-item',
@@ -6,10 +7,19 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./book-item.component.scss']
 })
 export class BookItemComponent implements OnInit {
+  @Input() book: BookModel;
+  @Output() add = new EventEmitter<BookModel>();
 
   constructor() { }
 
   ngOnInit(): void {
   }
 
+  dateToString(time: number): Date {
+    return new Date(time);
+  }
+
+  addBook(): void {
+    this.add.emit(this.book);
+  }
 }
